ecostock_backend/app_eco/requirements.txt
# Requirements pour app_eco
# =========================

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# OCR (optionnel mais recommande)
easyocr>=1.7.0
